server/requirements.txt
Flask>=2.3.0,<3.0.0
Flask-SQLAlchemy>=3.0.0,<4.0.0
python-dotenv>=1.0.0
